Renaming and updating to standards
