+++ src/Workflow/Attributes/Response/Content/WorkflowActionSubmissionJson.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
  *  @license    http://www.pimcore.org/license     GPLv3 and PCL
  */
 
-namespace Pimcore\Bundle\StudioBackendBundle\OpenApi\Attributes\Response\Content;
+namespace Pimcore\Bundle\StudioBackendBundle\Workflow\Attributes\Response\Content;
 
 use OpenApi\Attributes\JsonContent;
 use OpenApi\Attributes\Property;
