+++ src/Workflow/Attributes/Response/Property/WorkflowDetailsCollection.php
@@ -14,7 +14,7 @@ declare(strict_types=1);
  *  @license    http://www.pimcore.org/license     GPLv3 and PCL
  */
 
-namespace Pimcore\Bundle\StudioBackendBundle\OpenApi\Attributes\Response\Property;
+namespace Pimcore\Bundle\StudioBackendBundle\Workflow\Attributes\Response\Property;
 
 use OpenApi\Attributes\Items;
 use OpenApi\Attributes\Property;
